and some more renames
git-svn-id: http://svn.code.sf.net/p/freeglut/code/trunk/freeglut/freeglut@1112 7f0cb862-5218-0410-a997-914c9d46530a

+++ freeglut/src/mswin/fg_main_mswin.c
@@ -142,7 +142,7 @@ fg_time_t fgPlatformSystemTime ( void )
 
 void fgPlatformSleepForEvents( fg_time_t msec )
 {
-    MsgWaitForMultipleObjects( 0, NULL, FALSE, msec, QS_ALLINPUT );
+    MsgWaitForMultipleObjects( 0, NULL, FALSE, (DWORD) msec, QS_ALLINPUT );
 }
